update repo_clone_snapshot and python module
